requirements.txt
torch
torchvision
diffusers
transformers
accelerate
Pillow
flask
flask-cors
werkzeug
requests
